AvailableMarketsOptions selector
